+++ src/app/components/Routes/index.tsx
@@ -11,12 +11,14 @@ import styles from './styles.module.scss';
 
 const Home = lazy(() => import('../../screens/Dashboard'));
 const Login = lazy(() => import('../../screens/Login'));
+const Sidebar = lazy(() => import('../Sidebar'));
 
 function AppRoutes() {
   return (
     <ConnectedRouter history={history}>
-      <div className={styles.container}>
+      <div className={`row full-width ${styles.container}`}>
         <Suspense>
+          <Sidebar />
           <Switch>
             <AuthenticatedRoute isPublicRoute exact path={Routes.HOME} component={Home} />
             <AuthenticatedRoute isPublicRoute exact path={Routes.LOGIN} component={Login} />
