+++ src/app/components/Sidebar/constants.js
@@ -1,0 +1,22 @@
+export const CATEGORIES = [
+  {
+    id: 0,
+    name: 'BUTTONS'
+  },
+  {
+    id: 1,
+    name: 'CHECKBOXES'
+  },
+  {
+    id: 2,
+    name: 'INPUTS'
+  },
+  {
+    id: 3,
+    name: 'RADIO-BUTTONS'
+  },
+  {
+    id: 4,
+    name: 'SPINNERS'
+  }
+];
