+++ src/app/screens/Dashboard/screens/Home/constants.js
@@ -1,26 +1,3 @@
-export const CATEGORIES = [
-  {
-    id: 0,
-    name: 'BUTTONS'
-  },
-  {
-    id: 1,
-    name: 'CHECKBOXES'
-  },
-  {
-    id: 2,
-    name: 'INPUTS'
-  },
-  {
-    id: 3,
-    name: 'RADIO-BUTTONS'
-  },
-  {
-    id: 4,
-    name: 'SPINNERS'
-  }
-];
-
 export const COMPONENTS = [
   {
     name: 'button-1',
