+++ src/app/screens/Dashboard/screens/Home/styles.module.scss
@@ -1,0 +1,46 @@
+@import 'variables/colors';
+@import 'variables/sizes';
+
+.feed-content {
+  margin: 0 auto;
+  max-width: $content-max-width;
+  padding: 80px 20px 20px 80px;
+}
+
+.card-container {
+  display: grid;
+  grid-gap: 20px;
+  grid-template-columns: repeat(4, 1fr);
+  justify-items: center;
+}
+
+@media screen and (max-width: $big-screen) {
+
+  .card-container {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+
+@media screen and (max-width: $medium-screen) {
+
+  .card-container {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media screen and (max-width: $small-screen) {
+
+  .feed-content {
+    padding: 70px 20px 20px;
+  }
+
+  .card-container {
+    grid-template-columns: 1fr;
+    justify-items: center;
+  }
+
+  .title {
+    display: block;
+    text-align: center;
+  }
+}
